+++ beobench/__init__.py
@@ -1,8 +1,8 @@
 """Top-level package for beobench."""
 
-__author__ = """rdnfn"""
+__author__ = """Beobench authors"""
 __email__ = "-"
-__version__ = "0.4.3"
+__version__ = "0.4.4"
 
 from beobench.utils import restart
 from beobench.experiment.scheduler import run