+++ tests/conftest.py
@@ -4,6 +4,8 @@ import pytest
 
 import beobench.experiment.config_parser
 
+collect_ignore_glob = ["performance"]
+
 
 @pytest.fixture
 def run_config():